Refactor main script into modular server, engine, and UI
