+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "start": "node src/engine.mjs",
+    "start": "node src/server.mjs",
     "test": "node --test --test-concurrency=1"
   },
   "dependencies": {
