+++ src/ui/index.html
@@ -65,6 +65,6 @@
 </div>
 
 <script type="module">
-  import { boot } from '/web/preview.mjs';
+  import { boot } from './preview.mjs';
   boot();
 </script>
