+++ src/ui/readme.md
@@ -1,0 +1,7 @@
+# UI
+
+Browser interface providing live preview and controls.
+
+- `index.html` – control layout and canvas elements.
+- `preview.mjs` – draws preview scenes and syncs params over WebSocket.
+- `favicon.ico` – page icon.
